# Switch particle storage from AOS to SOA.