Not much refactoring, just transposing vector of tuples to tuple of vectors.

Overall the performance improved by ~5 seconds. Boundary condition function starts performing worse due to this change, but let it be a topic for a further investigation, since there are other problems with it and LTO, for example.

Before:
```
--------------------------------------------------------------------------------
abs. time [s]    rel. time [%]    calls [#]    section name
--------------------------------------------------------------------------------
    125.10847        100.00000            1    main
    115.56749         92.37383        34130    tit::EulerIntegrator::step()
     51.37776         41.06657         3413    tit::ParticleAdjacency::build()
     32.06365         25.62868        34130    tit::FluidEquations::compute_density()
     15.72200         12.56669        34130    tit::FluidEquations::compute_forces()
     14.04229         11.22409        34130    tit::FluidEquations::setup_boundary()
      3.07551          2.45827         3413    tit::ZCurveOrdering::ZCurveOrdering()
      0.93631          0.74840         3413    tit::Grid::Grid()
      0.00014          0.00011            1    tit::FluidEquations::init()
--------------------------------------------------------------------------------
```

After:
```
--------------------------------------------------------------------------------
abs. time [s]    rel. time [%]    calls [#]    section name
--------------------------------------------------------------------------------
    120.28590        100.00000            1    main
    110.72962         92.05536        34130    tit::EulerIntegrator::step()
     48.74499         40.52428         3413    tit::ParticleAdjacency::build() <-- faster by 3 seconds
     29.00364         24.11226        34130    tit::FluidEquations::compute_density() <-- faster by 3 seconds
     16.72462         13.90406        34130    tit::FluidEquations::setup_boundary() <-- slower by 1 second
     14.17562         11.78494        34130    tit::FluidEquations::compute_forces()
      2.93674          2.44146         3413    tit::ZCurveOrdering::ZCurveOrdering()
      0.84294          0.70078         3413    tit::Grid::Grid()
      0.00014          0.00011            1    tit::FluidEquations::init()
--------------------------------------------------------------------------------
```